# openapi-advance-app/src/main/resources/openapi/openapi-server.yml
openapi: 3.0.3
info:
  title: Swagger Openapi Server
  description: 'This is the Openapi Specification For Feign Server'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
tags:
- name: server
  description: all server apis

paths:
  /v1/server/customers:
    post:
      tags:
      - server
      summary: Create a customer
      description: create a customer
      operationId: createCustomer
      requestBody:
        description: request body
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerRequest'
        required: true
      responses:
        201:
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                format: uuid
    get:
      tags:
      - server
      summary: get customers
      description: get customers
      operationId: getCustomers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
components:
  schemas:
    CustomerRequest:
      type: object
      x-class-extra-annotation: "@lombok.Setter\n@lombok.Getter"
      x-internal-imports:
        - "import javax.validation.constraints.Max;"
        - "import javax.validation.constraints.Min;"
      properties:
        fullName:
          type: string
          example: Nguyen Minh Duc
          x-field-extra-annotation: "@Min(value = 1)"
        email:
          type: string
          format: email
          example: ducnguyen@gmail.com
        address:
          type: string
          x-field-extra-annotation: "@Max(value = 100)"
          example: 3/115 Binh Duong
        phone:
          type: string
          example: 0999123445
        gender:
          type: string
          enum: [M, F]
        dob:
          type: string
          format: date
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
          minLength: 16
        fullName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 3
        address:
          type: string
        phone:
          type: string
        gender:
          type: string
          enum: [ M, F ]
        dob:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time